[voq][chassis] Fix for issue in chassis redis server connection check
During database docker start, when the redis servers are started, ping
pong is done to check the server connection status. For server within
the card ping pong is done using unix socket. In VOQ chassis, to check
the connection to redis server in supervisor card, we need to use host
name and port instead of using unix socket.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>